Extract splitting logic and add tests
