+++ README.md
@@ -1,1 +1,11 @@
 # word-splitter
+
+## Running Tests
+
+Ensure you have Node.js installed, then execute:
+
+```bash
+node test.js
+```
+
+This runs a series of assertions against the `calculateParts` function.
