+++ index.html
@@ -20,6 +20,7 @@
             <input type="text" id="wordInput" placeholder="Enter a word">
         </div>
     </div>
+    <script src="splitter.js"></script>
     <script src="script.js"></script>
 </body>
 </html>
